Re worked where the ui window width and page width get set. These two parameters actually apply to the entire page so it made sense to move the redux action dispatch call out of the hydrograph specific code.
