+++ assets/src/scripts/store/daily-value-time-series.js
@@ -8,7 +8,7 @@ const INITIAL_DAILY_VALUE_TIME_SERIES_DATA= {
 /*
  * Synchronous Redux action to set the available dv time series
  * @param {Object} availableDVTimeSeries
- * @return {Function} which returns a Redux action
+ * @return {Object} Redux action
  */
 const setAvailableDVTimeSeries = function (availableTimeSeries) {
     return {
@@ -21,7 +21,7 @@ const setAvailableDVTimeSeries = function (availableTimeSeries) {
  * Synchronous Redux action to add the dv time series
  * @param {String} timeSeriesId
  * @param {Object} data
- * @return {Function} which returns a Redux action
+ * @return {Object} Redux action
  */
 const addDVTimeSeries = function(timeSeriesId, data) {
     return {
@@ -34,7 +34,7 @@ const addDVTimeSeries = function(timeSeriesId, data) {
 /*
  * Synchronous Redux Action to update the current time series id to be viewed
  * @param {String} timeSeriesId
- * @return {Function} which returns a Redux action
+ * @return {Object} Redux action
  */
 const setCurrentDVTimeSeriesId = function(timeSeriesId) {
     return {
@@ -46,7 +46,7 @@ const setCurrentDVTimeSeriesId = function(timeSeriesId) {
 /*
  * Synchronous Redux action to update the graph's cursor offset
  * @param {Number} cursorOffset - difference in epoch time from the cursor position to graph start time
- * @return {Function} which returns a Redux action
+ * @return {Object} Redux action
  */
 const setDVGraphCursorOffset = function(cursorOffset) {
     return {
@@ -131,7 +131,7 @@ export const dailyValueTimeSeriesDataReducer =
  * Synchronous Redux action to update the graph's brush offset
  * @param {Number} startBrushOffset - difference in epoch time from brush start to the displayed time series start time
  * @param {Number} endBrushOffset - difference in epoch time from displayed time series end to the end of the brush
- * @return {Function} which returns a Redux action
+ * @return {Object} Redux action
  */
 const setDVGraphBrushOffset = function(startBrushOffset, endBrushOffset) {
     return {
@@ -143,7 +143,7 @@ const setDVGraphBrushOffset = function(startBrushOffset, endBrushOffset) {
 
 /*
  * Synchronous Redux action to clear the graph's brush offset
- * @return {Function} which returns a Redux action
+ * @return {Object} Redux action
  */
 const clearDVGraphBrushOffset = function() {
     return {
